Fix VA getting stuck
When the virtual agent's target is taken, it will slow down to a stop
and pick a new target. When it picks a new target there's a possibility
of picking a relatively close target (maybe around 125 units). Via line
453, when the VA is close to the target it should only be decelerating.
I believe that when the VA gets to this point over a short distance, it
hasn't picked up enough speed and decelerates to a complete stop before
getting close enough to the target to tag it. In this case, we get no
movement. As a solution, I added an additional branch in update() so
that the VA will accelerate to the target anyways.